Extract log objects into log module
