+++ lib/server/startApp.js
@@ -2,13 +2,13 @@
 
 import create from './../../app/helpers/redux';
 import playerReducers from '../../app/modules/player/reducers';
-import consoleReducers from '../../app/modules/console/reducers';
+import logReducers from '../../app/modules/log/reducers';
 import { getPlayer } from '../../app/modules/player/actions/index';
-import { getLog } from '../../app/modules/console/actions/index';
+import { getLog } from '../../app/modules/log/actions/index';
 import render from './render';
 
 const startApp = function *() {
-  const reducers = Object.assign({}, { player: playerReducers }, { log: consoleReducers });
+  const reducers = Object.assign({}, { player: playerReducers }, { log: logReducers });
   const store = create(reducers);
   let html;
